% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{coord_radar-deprecated}
\alias{coord_radar-deprecated}
\alias{coord_radar}
\title{coord_radar (deprecated)}
\usage{
coord_radar(theta = "x", start = 0, direction = 1, clip = "on")
}
\arguments{
\item{theta}{variable to map angle to (\code{x} or \code{y})}

\item{start}{Offset of starting point from 12 o'clock in radians. Offset
is applied clockwise or anticlockwise depending on value of \code{direction}.}

\item{direction}{1, clockwise; -1, anticlockwise}

\item{clip}{Should drawing be clipped to the extent of the plot panel? A
setting of \code{"on"} (the default) means yes, and a setting of \code{"off"}
means no. For details, please see \code{\link[ggplot2:coord_cartesian]{coord_cartesian()}}.}
}
\description{
use \code{\link{coord_radial}}
}
\keyword{internal}
