% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation-function.R
\name{dot_product}
\alias{dot_product}
\alias{andrews}
\alias{legendre}
\title{Andrews plot}
\usage{
andrews(k = 4, length.out = 50 * (k - 1), ...)

legendre(k = 4, length.out = 50 * (k - 1), ...)
}
\arguments{
\item{k}{The number of dimensions}

\item{length.out}{The series length}

\item{...}{Other arguments passed on to methods. Mainly used for customized transformation function}
}
\value{
A list contains two contents
\enumerate{
  \item A series
  \item Transformed matrix
}
}
\description{
An Andrews plot or Andrews curve is a way to visualize structure in high-dimensional data
by defining a finite Fourier series.

Legendre polynomials are a system of complete and orthogonal polynomials.
}
\details{
In Andrews transformation, we project the data point onto the following k dimensional Fourier series,
(\bold{1/sqrt(2)}^T, sin(\bold{t}^T), cos(\bold{t}^T), sin(\bold{2t}^T), cos(\bold{2t}^T), ...)^T

P_n(\bold{x}) is polynomial of degreen n. Legendre polynomials are a list of polynomials match that

integral_{-1}^1 P_n(\bold{x})P_m(\bold{x}) dx = 0

where n is not equal to m
}
\references{
Andrews, David F. "Plots of high-dimensional data." \emph{Biometrics} (1972): 125-136.

Abramowitz, Milton, and Irene A. Stegun, eds. "Chapter 8"
\emph{Handbook of mathematical functions with formulas, graphs, and mathematical tables}.
Vol. 55. US Government printing office, 1948.
}
