% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-density-.R, R/geom-hist-.R,
%   R/geom-image-glyph.R, R/geom-polygon-glyph.R, R/geom-quantiles.R,
%   R/geom-serialaxes-density.R, R/geom-serialaxes-glyph.R,
%   R/geom-serialaxes-hist.R, R/geom-serialaxes-quantile.R, R/geom-serialaxes.R
\docType{data}
\name{Geom-ggproto}
\alias{Geom-ggproto}
\alias{GeomDensity_}
\alias{GeomBar_}
\alias{GeomImageGlyph}
\alias{GeomPolygonGlyph}
\alias{GeomQuantiles}
\alias{GeomSerialaxesDensity}
\alias{GeomSerialAxesGlyph}
\alias{GeomSerialaxesHist}
\alias{GeomSerialaxesQuantile}
\alias{GeomSerialaxes}
\title{Base Geom ggproto classes for ggplot2}
\format{
An object of class \code{GeomDensity_} (inherits from \code{GeomRibbon}, \code{Geom}, \code{ggproto}, \code{gg}) of length 6.

An object of class \code{GeomBar_} (inherits from \code{GeomBar}, \code{GeomRect}, \code{Geom}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{GeomImageGlyph} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 7.

An object of class \code{GeomPolygonGlyph} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 7.

An object of class \code{GeomQuantiles} (inherits from \code{GeomQuantile}, \code{GeomPath}, \code{Geom}, \code{ggproto}, \code{gg}) of length 1.

An object of class \code{GeomSerialaxesDensity} (inherits from \code{GeomDensity_}, \code{GeomRibbon}, \code{Geom}, \code{ggproto}, \code{gg}) of length 2.

An object of class \code{GeomSerialAxesGlyph} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 7.

An object of class \code{GeomSerialaxesHist} (inherits from \code{GeomBar_}, \code{GeomBar}, \code{GeomRect}, \code{Geom}, \code{ggproto}, \code{gg}) of length 2.

An object of class \code{GeomSerialaxesQuantile} (inherits from \code{GeomPath}, \code{Geom}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{GeomSerialaxes} (inherits from \code{GeomPath}, \code{Geom}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
GeomDensity_

GeomBar_

GeomImageGlyph

GeomPolygonGlyph

GeomQuantiles

GeomSerialaxesDensity

GeomSerialAxesGlyph

GeomSerialaxesHist

GeomSerialaxesQuantile

GeomSerialaxes
}
\description{
All \code{geom_} functions (like \code{geom_point}) return a layer that contains a \code{Geom}
object (like \code{GeomPoint}). The \code{Geom} object is responsible for rendering the data in the plot.
Each of the \code{Geom} objects is a \code{ggproto} object,
descended from the top-level Geom, and each implements various methods and fields.
Compared to \code{Stat} and \code{Position}, \code{Geom} is a little different
because the execution of the setup and compute functions is split up.
setup_data runs before position adjustments, and \code{draw_layer}
is not run until render time, much later. This means there is no \code{setup_params}
because it's hard to communicate the changes.
}
\keyword{datasets}
