% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-image-glyph.R
\name{geom_image_glyph}
\alias{geom_image_glyph}
\title{Add image glyphs on scatter plot}
\usage{
geom_image_glyph(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  images,
  imagewidth = 1.2,
  imageheight = 0.9,
  interpolate = TRUE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, either as a \code{ggproto} \code{Geom} subclass or as a string naming the
stat stripped of the \code{stat_} prefix (e.g. \code{"count"} rather than
\code{"stat_count"})}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{images}{a list of images (a raster object, bitmap image). If not provided, a point visual (\code{geom_point()}) will be displayed.}

\item{imagewidth}{Numerical; width of image}

\item{imageheight}{Numerical; height of image}

\item{interpolate}{A logical value indicating whether to linearly interpolate the image (the alternative is to use nearest-neighbour interpolation,
which gives a more blocky result). See \code{\link{rasterGrob}}.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\value{
a \code{geom} layer
}
\description{
Each point glyph can be an image (png, jpeg, etc) object.
}
\section{Aesthetics}{

geom_..._glyph() understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item{\strong{x}}
\item{\strong{y}}
\item{alpha}
\item{colour}
\item{fill}
\item{group}
\item{size}
\item{linetype}
\item{shape}
\item{stroke}
}

The size unit is \code{cm}

Note that the shape and stroke do not have real meanings unless the essential
argument \code{images} is missing. If so, a point visual will be displayed with
corresponding shape and stroke.
}

\examples{
\donttest{
# image glyph
if(require("png")) {
img_path <- list.files(file.path(find.package(package = 'ggmulti'),
                                 "images"),
                       full.names = TRUE)
Raptors <- png::readPNG(img_path[2L])
Warriors <- png::readPNG(img_path[3L])

pg <- ggplot(data = data.frame(x = 1:2, y = rep(1, 2)),
       mapping = aes(x = x, y = y)) +
  geom_image_glyph(images = list(Raptors,
                                 Warriors),
                   imagewidth = rep(1.2, 2),
                   imageheight = c(0.9, 1.2)) +
  coord_cartesian(xlim = extendrange(c(1,2)))
pg
# query the images (a numerical array)
build <- ggplot2::ggplot_build(pg)
# `imageRaptors` and `imageWarriors` are three dimensional
# arrays (third dimension specifying the plane)
imageRaptors <- build$data[[1]]$images[[1]]
imageWarriors <- build$data[[1]]$images[[2]]

if(require("grid")) {
grid.newpage()
grid.raster(imageRaptors)
grid.newpage()
grid.raster(imageWarriors)
}

# THIS IS SLOW
mercLogo <- png::readPNG(img_path[1L])

p <- ggplot(mapping = aes(x = hp, y = mpg)) +
       geom_point(
         data = mtcars[!grepl("Merc", rownames(mtcars)), ],
         color = "skyblue") +
       geom_image_glyph(
         data = mtcars[grepl("Merc", rownames(mtcars)), ],
         images = mercLogo,
         imagewidth = 1.5
       )
p

}
}
}
\seealso{
\code{\link{geom_serialaxes_glyph}}, \code{\link{geom_polygon_glyph}}
}
