% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmulti-package.R
\docType{package}
\name{ggmulti-package}
\alias{ggmulti-package}
\alias{_PACKAGE}
\title{ggmulti: High Dimensional Data Visualization}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

It provides materials (i.e. 'serial axes' objects, Andrew's plot, various glyphs for scatter plot) to visualize high dimensional data.
}
\author{
\strong{Maintainer}: Zehao Xu \email{z267xu@uwaterloo.ca}

Authors:
\itemize{
  \item R. Wayne Oldford \email{rwoldford@uwaterloo.ca}
}

}
\keyword{internal}
