% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon-chars.R
\docType{data}
\name{polygon_glyph}
\alias{polygon_glyph}
\alias{x_star}
\alias{y_star}
\alias{x_cross}
\alias{y_cross}
\alias{x_hexagon}
\alias{y_hexagon}
\alias{x_airplane}
\alias{y_airplane}
\alias{x_maple}
\alias{y_maple}
\title{Polygon glyph coordinates}
\format{
An object of class \code{numeric} of length 10.

An object of class \code{numeric} of length 10.

An object of class \code{numeric} of length 12.

An object of class \code{numeric} of length 12.

An object of class \code{numeric} of length 6.

An object of class \code{numeric} of length 6.

An object of class \code{numeric} of length 32.

An object of class \code{numeric} of length 32.

An object of class \code{numeric} of length 26.

An object of class \code{numeric} of length 26.
}
\usage{
x_star

y_star

x_cross

y_cross

x_hexagon

y_hexagon

x_airplane

y_airplane

x_maple

y_maple
}
\description{
polygon coordinates scaled to (0, 1)
}
\examples{
if(requireNamespace("grid")) {
  library(grid)
  grid.newpage()
  grid.polygon(x=(x_star + 1)/2,
               y=(y_star + 1)/2)
  grid.newpage()
  grid.polygon(x=(x_cross + 1)/2,
               y=(y_cross + 1)/2)
  grid.newpage()
  grid.polygon(x=(x_hexagon + 1)/2,
               y=(y_hexagon + 1)/2)
  grid.newpage()
  grid.polygon(x=(x_airplane + 1)/2,
               y=(y_airplane + 1)/2)
  grid.newpage()
  grid.polygon(x=(x_maple + 1)/2,
               y=(y_maple + 1)/2)
}
}
\seealso{
\link{geom_polygon_glyph}
}
\keyword{datasets}
