% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-dodge-.R, R/position-identity-.R,
%   R/position-stack-.R
\docType{data}
\name{Position-ggproto}
\alias{Position-ggproto}
\alias{PositionDodge_}
\alias{PositionDodge2_}
\alias{PositionIdentity_}
\alias{PositionStack_}
\alias{PositionFill_}
\title{Base Position ggproto classes for ggplot2}
\format{
An object of class \code{PositionDodge_} (inherits from \code{PositionDodge}, \code{Position}, \code{ggproto}, \code{gg}) of length 2.

An object of class \code{PositionDodge2_} (inherits from \code{PositionDodge2}, \code{PositionDodge}, \code{Position}, \code{ggproto}, \code{gg}) of length 2.

An object of class \code{PositionIdentity_} (inherits from \code{PositionIdentity}, \code{Position}, \code{ggproto}, \code{gg}) of length 3.

An object of class \code{PositionStack_} (inherits from \code{PositionStack}, \code{Position}, \code{ggproto}, \code{gg}) of length 3.

An object of class \code{PositionFill_} (inherits from \code{PositionStack_}, \code{PositionStack}, \code{Position}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
PositionDodge_

PositionDodge2_

PositionIdentity_

PositionStack_

PositionFill_
}
\description{
All \code{position_} functions (like \code{position_dodge})
return a \code{Position} object (like \code{PositionDodge}).
The \code{Position} object is responsible for adjusting the position of overlapping geoms.
The way that the \code{position_} functions work is slightly different from the
\code{geom_} and \code{stat_} functions, because a \code{position_} function
actually "instantiates" the \code{Position} object by creating a descendant,
and returns that.
Each of the \code{Position} objects is a \code{ggproto} object,
descended from the top-level \code{Position}.
}
\keyword{datasets}
