% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-identity-.R
\name{position_identity_}
\alias{position_identity_}
\title{Don't adjust position}
\usage{
position_identity_()
}
\description{
Don't adjust position
}
\seealso{
Other position adjustments for multiple locations:
\code{\link{position_stack_}}, \code{\link{position_fill_}},
\code{\link{position_dodge_}}, \code{\link{position_dodge2_}}
}
