% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-density-.R, R/stat-hist-.R,
%   R/stat-serialaxes-density.R, R/stat-serialaxes-hist.R, R/stat-serialaxes.R
\docType{data}
\name{Stat-ggproto}
\alias{Stat-ggproto}
\alias{StatDensity_}
\alias{StatHist_}
\alias{StatBin_}
\alias{StatCount_}
\alias{StatSerialaxesDensity}
\alias{StatSerialaxesHist}
\alias{StatSerialaxes}
\alias{StatDotProduct}
\title{Base Stat ggproto classes for ggplot2}
\format{
An object of class \code{StatDensity_} (inherits from \code{StatDensity}, \code{Stat}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{StatHist_} (inherits from \code{StatBin}, \code{Stat}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{StatBin_} (inherits from \code{StatHist_}, \code{StatBin}, \code{Stat}, \code{ggproto}, \code{gg}) of length 2.

An object of class \code{StatCount_} (inherits from \code{StatHist_}, \code{StatBin}, \code{Stat}, \code{ggproto}, \code{gg}) of length 2.

An object of class \code{StatSerialaxesDensity} (inherits from \code{StatDensity}, \code{Stat}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{StatSerialaxesHist} (inherits from \code{StatBin}, \code{Stat}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{StatSerialaxes} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 6.

An object of class \code{StatDotProduct} (inherits from \code{StatSerialaxes}, \code{Stat}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
StatDensity_

StatHist_

StatBin_

StatCount_

StatSerialaxesDensity

StatSerialaxesHist

StatSerialaxes

StatDotProduct
}
\description{
All \code{stat_} functions (like \code{stat_bin()})
return a layer that contains a \code{Stat} object (like \code{StatBin}).
The  \code{Stat} object is responsible for rendering the data in the plot.
Each of the \code{Stat} objects is a \code{ggproto} object,
descended from the top-level \code{Stat}, and each implements various methods and fields.
}
\keyword{datasets}
