% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-serialaxes-layers.R
\name{add_serialaxes_layers}
\alias{add_serialaxes_layers}
\title{Layers for serial axes coordinate}
\usage{
add_serialaxes_layers(layer, plot, object, axes)
}
\arguments{
\item{layer}{a layer object}

\item{plot}{a \code{ggplot} object}

\item{object}{some parameters used to modify this serial axes \code{ggplot} object (i.e. \code{axes.sequence}, ...)}

\item{axes}{canvas sequence axes}
}
\description{
Project the regular \code{geom} layers onto the serial axes coordinate.
}
\details{
The class is determined by layers you add. For example, you want to add a boxplot layer
on serial axes coordinate.
By the ggplot syntax, it should be \code{ggplot(data, mapping) + geom_boxplot() + coord_serialaxes()}
To make it work, object \code{add_serialaxes_layers.GeomBoxplot} must be created. In this function,
some computations will be applied.
}
